BetaTrax/pyproject.toml
[project]
name = "BetaTraxService-GroupK"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Jchan", email = "u3621801@connect.hku.hk"},
    {name = "King", email = "kwhk0626@connect.hku.hk"},
    {name = "Nicholas", email = "nickyl12@connect.hku.hk"},
    {name = "Joe", email = "sangho12@connect.hku.hk"}
]
dependencies = ["Django==6.0.*", "djangorestframework>=3.17.1"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
